tag接口 shopID过滤

+++ models/tag.go
@@ -10,7 +10,7 @@ import (
 type Tag struct {
 	// 商品标签通常只有几十个，数据量小，不使用雪花ID
 	ID          int       `gorm:"primarykey" json:"id"`
-	ShopID      int       `gorm:"index;not null" json:"shop_id"` // 新增店铺ID
+	ShopID      uint64    `gorm:"index;not null" json:"shop_id"` // 新增店铺ID
 	Name        string    `gorm:"size:50;not null;uniqueIndex" json:"name"`
 	Description string    `gorm:"size:200" json:"description"`
 	CreatedAt   time.Time `json:"created_at"`
